+++ CMakeLists.txt
@@ -11,7 +11,7 @@ find_package( OpenCV REQUIRED )
 find_package( ROOT 6.18 CONFIG REQUIRED)
 set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-ADD_LIBRARY( featurereco_lib Configuration.cpp MedianTextReader.cpp)
+ADD_LIBRARY( featurereco_lib Configuration.cpp MedianTextReader.cpp featureFunctions.cpp)
 include_directories( ${OpenCV_INCLUDE_DIRS} ${ROOT_INCLUDE_DIRS})
 add_executable( FindBoltLocations FindBoltLocations.cpp )
 target_link_libraries( FindBoltLocations featurereco_lib ${OpenCV_LIBS} ${ROOT_LIBS} ROOT::Core ROOT::Hist)
